Error on confirm when using set as default feature + SSC (#4599) GIT_VALID_PII_OVERRIDE
<!-- Simple summary of what was changed. -->
The set as default feature does not support deferred server-side
confirmation, so we crash the app in test mode if a merchant tries to
confirm with deferred ssc and set as default.
<!-- Why are you making this change? If it's for fixing a bug, if
possible, please include a code snippet or example project that
demonstrates the issue. -->
[MOBILESDK-3050](https://jira.corp.stripe.com/browse/MOBILESDK-3050)
<!-- How was the code tested? Be as specific as possible. -->

<!-- Is this a notable change that affects users? If so, add a line to
`CHANGELOG.md` and prefix the line with one of the following:
    - [Added] for new features.
    - [Changed] for changes in existing functionality.
    - [Deprecated] for soon-to-be removed features.
    - [Removed] for now removed features.
    - [Fixed] for any bug fixes.
    - [Security] in case of vulnerabilities.
-->