Replace `las_client_secret` in logs with `las_id` (#4598)
## Summary

Instead of logging the client secret, we should now log the link account
session client secret. This replaces those parameters, passes along the
link account session ID when it is available (i.e. after we have
received the manifest).

## Motivation

This is in the spirit of least-privilege.

## Testing

N/a

## Changelog

N/a