Update US Bank Account payment method parsing (#4603)
## Summary
<!-- Simple summary of what was changed. -->

This pull request fixes an issue in the parsing logic of
`STPPaymentMethodUSBankAccount`, which used the `linked_account` key
instead of `financial_connections_account`.

We’re changing the key behind the scenes, but keeping the variable name
to avoid a breaking change. We’ll rename the field in the next major
version bump.

## Motivation
<!-- Why are you making this change? If it's for fixing a bug, if
possible, please include a code snippet or example project that
demonstrates the issue. -->

## Testing
<!-- How was the code tested? Be as specific as possible. -->

## Changelog
<!-- Is this a notable change that affects users? If so, add a line to
`CHANGELOG.md` and prefix the line with one of the following:
    - [Added] for new features.
    - [Changed] for changes in existing functionality.
    - [Deprecated] for soon-to-be removed features.
    - [Removed] for now removed features.
    - [Fixed] for any bug fixes.
    - [Security] in case of vulnerabilities.
-->