Add error screen for no connectivity situations. (#4601)
## Summary
This change adds an error screen if the component completely fails to
load. Otherwise after the component has loaded it is on the component to
display the error.

This is also the first use case for localization in the connect sdk, so
it also adds all of the needed localization files.

## Demo
![Simulator Screenshot - iPhone SE (3rd generation) - 2025-02-25 at 16
48
25](https://github.com/user-attachments/assets/03004fdb-f82a-4c2b-8a92-1cf581773415)
![Simulator Screenshot - iPad (10th generation) - 2025-02-25 at 16 45
37](https://github.com/user-attachments/assets/af524aeb-4a0f-4d1b-ad8d-0d4a8e6b07df)
![Simulator Screenshot - iPad (10th generation) - 2025-02-25 at 16 45
24](https://github.com/user-attachments/assets/4e77eaed-c3ef-4130-8a1c-df4a7bbfedfa)