CustomerSheet does not preselect unless it is default (#4607)
## Summary
<!-- Simple summary of what was changed. -->
CustomerPaymentOption.selectedPaymentMethod returns the default or first
payment method if it's .paymentSheet and only the default otherwise

Nothing in CustomerSheet will be selected unless it is the default.

## Motivation
<!-- Why are you making this change? If it's for fixing a bug, if
possible, please include a code snippet or example project that
demonstrates the issue. -->
Bug bash bug
## Testing
<!-- How was the code tested? Be as specific as possible. -->

## Changelog
<!-- Is this a notable change that affects users? If so, add a line to
`CHANGELOG.md` and prefix the line with one of the following:
    - [Added] for new features.
    - [Changed] for changes in existing functionality.
    - [Deprecated] for soon-to-be removed features.
    - [Removed] for now removed features.
    - [Fixed] for any bug fixes.
    - [Security] in case of vulnerabilities.
-->