Refactors TextFieldElement hasDisabledAppearance to allow for individual control of each field (#4608)
## Summary
Pre-work required for updating a payment method

## Motivation
We were previously setting the disabled appearance to the entire
section, but we need to be able to control individual fields.

## Testing
Manually and relying on existing snapshot tests -- though, there are
some severe limitations with snapshot tests. Darkmode tests no longer
show the mask being applied to each field, unfortunately.

## Changelog
<!-- Is this a notable change that affects users? If so, add a line to
`CHANGELOG.md` and prefix the line with one of the following:
    - [Added] for new features.
    - [Changed] for changes in existing functionality.
    - [Deprecated] for soon-to-be removed features.
    - [Removed] for now removed features.
    - [Fixed] for any bug fixes.
    - [Security] in case of vulnerabilities.
-->