Fix default badge bug (#4606)
## Summary
<!-- Simple summary of what was changed. -->
Replace row button for the previous default payment method if it wasn't
also previously selected
## Motivation
<!-- Why are you making this change? If it's for fixing a bug, if
possible, please include a code snippet or example project that
demonstrates the issue. -->
Bug bash bug
Default badge showed up on more than one row button because we were only
replacing the previously selected row button, but the previously
selected row button is not necessarily the default, so there was a
scenario where the default badge did not get cleared from the previous
default.
## Testing
<!-- How was the code tested? Be as specific as possible. -->

## Changelog
<!-- Is this a notable change that affects users? If so, add a line to
`CHANGELOG.md` and prefix the line with one of the following:
    - [Added] for new features.
    - [Changed] for changes in existing functionality.
    - [Deprecated] for soon-to-be removed features.
    - [Removed] for now removed features.
    - [Fixed] for any bug fixes.
    - [Security] in case of vulnerabilities.
-->