Move consumer session creation before verification screens (#4632)
## Summary
<!-- Simple summary of what was changed. -->

This pull request aligns the Financial Connections iOS SDK with Android
and Web, where we create a consumer session before we display any
verification screens. This removes a decent amount of error handling
code (for very unlikely errors) and will make implementing
https://github.com/stripe/stripe-ios/pull/4272 correctly easier.

## Motivation
<!-- Why are you making this change? If it's for fixing a bug, if
possible, please include a code snippet or example project that
demonstrates the issue. -->

## Testing
<!-- How was the code tested? Be as specific as possible. -->

## Changelog
<!-- Is this a notable change that affects users? If so, add a line to
`CHANGELOG.md` and prefix the line with one of the following:
    - [Added] for new features.
    - [Changed] for changes in existing functionality.
    - [Deprecated] for soon-to-be removed features.
    - [Removed] for now removed features.
    - [Fixed] for any bug fixes.
    - [Security] in case of vulnerabilities.
-->

---------

Co-authored-by: Mat Schmid <mats@stripe.com>