show disabled checkbox when it is the default pm (#4641)
## Summary
<!-- Simple summary of what was changed. -->
Always display the default checkbox on the update payment method screen.
Disable it if it is already the default and change the display text to
"Default payment method"
## Motivation
<!-- Why are you making this change? If it's for fixing a bug, if
possible, please include a code snippet or example project that
demonstrates the issue. -->
https://jira.corp.stripe.com/browse/MOBILESDK-3260
## Testing
<!-- How was the code tested? Be as specific as possible. -->
Updated UI tests, snapshot tests
## Changelog
<!-- Is this a notable change that affects users? If so, add a line to
`CHANGELOG.md` and prefix the line with one of the following:
    - [Added] for new features.
    - [Changed] for changes in existing functionality.
    - [Deprecated] for soon-to-be removed features.
    - [Removed] for now removed features.
    - [Fixed] for any bug fixes.
    - [Security] in case of vulnerabilities.
-->