Hide default checkbox and auto-set if first (#4643)
## Summary
<!-- Simple summary of what was changed. -->
If a customer has no saved payment methods and the default feature is
enabled, hide the default checkbox and automatically set new card or US
bank account as the default.
## Motivation
<!-- Why are you making this change? If it's for fixing a bug, if
possible, please include a code snippet or example project that
demonstrates the issue. -->
Bug bash suggestion
## Testing
<!-- How was the code tested? Be as specific as possible. -->
Updated UI test
## Changelog
<!-- Is this a notable change that affects users? If so, add a line to
`CHANGELOG.md` and prefix the line with one of the following:
    - [Added] for new features.
    - [Changed] for changes in existing functionality.
    - [Deprecated] for soon-to-be removed features.
    - [Removed] for now removed features.
    - [Fixed] for any bug fixes.
    - [Security] in case of vulnerabilities.
-->