Fix edit button behavior (#4667)
## Summary
<!-- Simple summary of what was changed. -->
Updated the check for whether or not the edit button should be shown to
include the default feature check. This oversight was in horizontal mode
only
## Motivation
<!-- Why are you making this change? If it's for fixing a bug, if
possible, please include a code snippet or example project that
demonstrates the issue. -->
Bug: disable removing payment methods, enable default
Expected behavior: you can manage the payment methods by setting default
Actual behavior: you cannot edit payment methods
## Testing
<!-- How was the code tested? Be as specific as possible. -->

- Manually

## Changelog
<!-- Is this a notable change that affects users? If so, add a line to
`CHANGELOG.md` and prefix the line with one of the following:
    - [Added] for new features.
    - [Changed] for changes in existing functionality.
    - [Deprecated] for soon-to-be removed features.
    - [Removed] for now removed features.
    - [Fixed] for any bug fixes.
    - [Security] in case of vulnerabilities.
-->
N/A