[FCLite] Add interface, models, and API client (#4685)
```
  ______    _____   _       _  _          
 |  ____|  / ____| | |     (_)| |         
 | |__    | |      | |      _ | |_   ___  
 |  __|   | |      | |     | || __| / _ \ 
 | |      | |____  | |____ | || |_ |  __/ 
 |_|       \_____| |______||_| \__| \___| 
```

Context:
https://docs.google.com/document/d/1Hb-HNFohgPWHm7951mr_2kOKZZ6xCgqRVp5zabAoXbg/edit?tab=t.0#heading=h.f6cv5x7p638o

## Summary

This adds the initial interface for Financial Connections Lite, along
with some models and an API client it will use. These will all be used
in subsequent PRs.

## Motivation

Bank payments 📈

## Testing

N/a - will add tests in subsequent PRs, along with a ton of manual
testing before anything is release.

## Changelog

N/a