[FCLite] Implement auth flow and container view controllers (#4687)
> [!NOTE]  
> Sorry for the large PR! I've broken this up into (hopefully)
reviewable commits!

## Summary

This implements most of the FC Lite flow. Here are the different working
pieces:
- `FinancialConnectionsLite`: Handles modal presentation logic. Retains
an instance of itself while the flow is active.
- `FCLiteContainerViewController`: Responsible for fetching the
manifest, showing an error screen if needed, and fetching the session
receipt at the end of the flow.
- `FCLiteAuthFlowViewController`: Responsible for managing the webview,
handling OAuth secure browser sessions, and app to app.

Here's the general flow of things:
1. A `FinancialConnectionsLite` instance is made by the host app with a
client secret and an optional return URL.
2. `.present` is called, which creates an
`FCLiteContainerViewController`, wraps it in an
`FCLiteModalPresentationWrapper`, then presents it modally above the
given view controller.
3. `FCLiteContainerViewController` shows a loading spinner while it
fetches the manifest. Once a manifest is receives, passes it to
`FCLiteAuthFlowViewController` to kick off the webview. If the request
fails, shows a simple error view with a retry button.
4. `FCLiteAuthFlowViewController` creates a `WKWebView` with the given
`manifest.hostAuthUrl`. For instant debits, we append a few query
parameters to the end of this URL. A progress bar is shown at the top of
the webview while it loads.
5. We observe redirects that happen within the webview. 
a. If we detect redirects to `manifest.successUrl` or
`manifest.cancelUrl`, we start the completion process.
b. If we detect attempts to open in a new window, we start the OAuth
process of either launching the app-to-app flow, or launching a secure
webview session.
7. Once the flow completes, we divert back to the
`FCLiteContainerViewController` to either fetch the session for FC, or
to extract the payment method object from the success URL for Instant
Debits.
8. Dismiss `FinancialConnectionsLite` and complete with the result.

Next, we'll be adding support to launch FC Lite from MPE when the native
SDK is not detected.

## Motivation

Bank payments 📈

## Testing

Some unit tests added around the `hostedAuthUrl` modification, and tons
of manual testing. In a subsequent PR we'll be adding end to end tests.

FC flow on iPhone:


https://github.com/user-attachments/assets/052ed9ef-2533-495b-ba46-7435d83bbb4f

Instant debits flow on iPhone:


https://github.com/user-attachments/assets/1b7013ac-d26a-427b-adef-a38b4b699220

iPad:


https://github.com/user-attachments/assets/5fdc824d-c2de-41ee-a82f-6ddb1364316e

iPad portrait:

<img width=50%
src="https://github.com/user-attachments/assets/d9e9adf8-06e0-4c93-b9eb-c6b9b00036b4">

Error screen:


https://github.com/user-attachments/assets/9b417315-b0c6-41e4-b167-39c9166e432b


## Changelog

N/a - this is only accessible through the FC Example app.