Filter out hidden elements in FormElement (#4704)
## Summary
<!-- Simple summary of what was changed. -->
HiddenElements within FormElements would take up space despite having a
.zero frame because of custom spacing applied between all elements in
the FormElement. Changed it so that the ViewModel filters out the
HiddenElements
## Motivation
<!-- Why are you making this change? If it's for fixing a bug, if
possible, please include a code snippet or example project that
demonstrates the issue. -->
[MOBILESDK-3359](https://jira.corp.stripe.com/browse/MOBILESDK-3359)
## Testing
<!-- How was the code tested? Be as specific as possible. -->

- Manually
When default feature is enabled and no previous saved pms, tapping the
save checkbox previously added a little bit of space where the hidden
checkbox would be. Now, there is no visual impact

## Changelog
<!-- Is this a notable change that affects users? If so, add a line to
`CHANGELOG.md` and prefix the line with one of the following:
    - [Added] for new features.
    - [Changed] for changes in existing functionality.
    - [Deprecated] for soon-to-be removed features.
    - [Removed] for now removed features.
    - [Fixed] for any bug fixes.
    - [Security] in case of vulnerabilities.
-->
N/A