“Set as default” checkbox preserves state (#4707)
## Summary
<!-- Simple summary of what was changed. -->
"Set as default checkbox" checks previous customer input to determine if
the checkbox should be checked
## Motivation
<!-- Why are you making this change? If it's for fixing a bug, if
possible, please include a code snippet or example project that
demonstrates the issue. -->
[MOBILESDK-3351](https://jira.corp.stripe.com/browse/MOBILESDK-3351)
## Testing
<!-- How was the code tested? Be as specific as possible. -->

- Manually
paymentMethodSetAsDefault enabled
In embedded continue mode, enter a card number, select "save" and "set
as default" checkboxes and continue. Change the mode (PI, PI+SFU, SI),
tap "change", and note the checkbox state

## Changelog
<!-- Is this a notable change that affects users? If so, add a line to
`CHANGELOG.md` and prefix the line with one of the following:
    - [Added] for new features.
    - [Changed] for changes in existing functionality.
    - [Deprecated] for soon-to-be removed features.
    - [Removed] for now removed features.
    - [Fixed] for any bug fixes.
    - [Security] in case of vulnerabilities.
-->
N/A