Add optimistic delete for CustomerSheet (#4738)
## Summary
Optimistically delete pms in customer sheet. Doing so leads to a
situation where we still see the pm in the list. This seems to be the
approach we've taken for payment sheet, so seems fine for customersheet.

## Motivation
Seeing some analytics that are warning us about the code.

## Testing
manual and relying on existing code.

## Changelog
<!-- Is this a notable change that affects users? If so, add a line to
`CHANGELOG.md` and prefix the line with one of the following:
    - [Added] for new features.
    - [Changed] for changes in existing functionality.
    - [Deprecated] for soon-to-be removed features.
    - [Removed] for now removed features.
    - [Fixed] for any bug fixes.
    - [Security] in case of vulnerabilities.
-->