[FCLite] Make FC Lite available in MPE (#4712)
## Summary

This integrates FC Lite into MPE. The main idea here is we first check
if the full SDK is available, use that if so, and otherwise fallback on
FC Lite.

> [!IMPORTANT]  
>This PR does not release FC Lite. There is a guardrail in place
(`fcLiteFeatureEnabled`) that prevents it from being accessed, and this
is only currently enabled by our example apps.

## Motivation
<!-- Why are you making this change? If it's for fixing a bug, if
possible, please include a code snippet or example project that
demonstrates the issue. -->

## Testing

Manual testing done! Here's the Payment Sheet example app: 


https://github.com/user-attachments/assets/d9c7f0ef-fcff-48ea-b5fd-ed89b64f36fb

And the Financial Connections example app:


https://github.com/user-attachments/assets/e47df2f8-362e-49d0-a7db-0b5f401d8c8b

## Changelog

N/a