Fix CardScan animation (#4714)
## Summary
This "Scan Card" animation has been bugging me: It overlaps the other
fields as it animates out.

* To fix this, I pinned it to the top of a container view, unpinned the
bottom, and enabled clipsToBounds. It'll now stay stuck to the top of
the container while the container's size animates down to 0.
* I also moved the disabling of the captureSession to *after* the
animation kicks off to make closing the view a little more responsive,
as disconnecting the camera input blocks the main thread for a bit.
* Also switched the animation to use a basic spring.

| Before | After |
| ------ | ----- |
| ![CleanShot 2025-03-25 at 14 41
10](https://github.com/user-attachments/assets/f1d58a56-378a-4ef3-b59e-8c9b07273d8d)
| ![CleanShot 2025-03-25 at 14 39
29](https://github.com/user-attachments/assets/1724d3b7-f13f-41fe-92b4-18100a8e836a)
|

## Motivation
https://jira.corp.stripe.com/browse/MOBILESDK-1302

## Testing
Tested in simulator and on an iPhone 16 Pro

## Changelog
Not significant enough