Return type `card` in FlowController when using Link inline signup (#4728)
## Summary
<!-- Simple summary of what was changed. -->

This pull request fixes an issue in FlowController where paying with a
card while also signing up for Link would return a
`PaymentOptionDisplayData` with the wrong payment method icon (the Link
icon instead of the card brand icon).

## Motivation
<!-- Why are you making this change? If it's for fixing a bug, if
possible, please include a code snippet or example project that
demonstrates the issue. -->

[LINK_MOBILE-26](https://jira.corp.stripe.com/browse/LINK_MOBILE-26)

## Testing
<!-- How was the code tested? Be as specific as possible. -->

Added an end-to-end test.

## Changelog
<!-- Is this a notable change that affects users? If so, add a line to
`CHANGELOG.md` and prefix the line with one of the following:
    - [Added] for new features.
    - [Changed] for changes in existing functionality.
    - [Deprecated] for soon-to-be removed features.
    - [Removed] for now removed features.
    - [Fixed] for any bug fixes.
    - [Security] in case of vulnerabilities.
-->