Improve expiration date scanning (#4719)
## Summary
Due to improvements in iOS' OCR and changes to modern cards, we're often
seeing strings like `Valid Thru 01/30 Security Code 123`, which are not
parsed by the existing logic. To catch these, we'll add another check to
look for dates in a "00/00" format.

Before: 5/14 test cards successfully populated an expiration date on
scan
After: 14/14 test cards succeeded. Scanning was marginally faster, as we
don't have to wait for the timeout

## Motivation
https://jira.corp.stripe.com/browse/MOBILESDK-3369

## Testing
Tested on iPhone 16 Pro, iOS 18.3.2

## Changelog
Adding