Hide Link logo in signup screen text field (#4761)
## Summary
<!-- Simple summary of what was changed. -->

This pull request removes the Link logo from the email text field in the
Link signup screen. We do this since we already show the logo at the top
of the screen.

## Motivation
<!-- Why are you making this change? If it's for fixing a bug, if
possible, please include a code snippet or example project that
demonstrates the issue. -->

[LINK_MOBILE-91](https://jira.corp.stripe.com/browse/LINK_MOBILE-91)

## Testing
<!-- How was the code tested? Be as specific as possible. -->

## Changelog
<!-- Is this a notable change that affects users? If so, add a line to
`CHANGELOG.md` and prefix the line with one of the following:
    - [Added] for new features.
    - [Changed] for changes in existing functionality.
    - [Deprecated] for soon-to-be removed features.
    - [Removed] for now removed features.
    - [Fixed] for any bug fixes.
    - [Security] in case of vulnerabilities.
-->