[FCLite] Add UI test (#4741)
## Summary

Adds a UI test for the FC Lite flow integration in the PaymentSheet
Example app. The test is intentionally kept fairly barebones. By nature
of this flow being a webview, the content is prone to changes, and we
don't want to introduce flaky tests.


https://github.com/user-attachments/assets/ed83ca0a-8c9c-4e53-9017-6067511a11d1

## Motivation

Ensure the flow opens and closes.

## Testing

✅ 

## Changelog

N/a