Support BDCC in native Link (#4692)
## Summary
<!-- Simple summary of what was changed. -->

This pull request allows Link to be enabled even if the merchant is
providing a custom billing details collection configuration (BDCC).

We pass the BDCC along to the native Link experience and collect the
required billing details there:
- If a consumer pays with a new payment method, we collect the fields
right there.
- If a consumer pays with an existing payment method, we evaluate
whether they satisfy the BDCC. We try to prefill empty fields with any
information we have through the Link account, such as the phone number.
If not, we present a screen for them to “confirm payment details”.

## Motivation
<!-- Why are you making this change? If it's for fixing a bug, if
possible, please include a code snippet or example project that
demonstrates the issue. -->

## Testing
<!-- How was the code tested? Be as specific as possible. -->

End-to-end tests for new and existing consumers, in both payment method
mode and passthrough mode.

## Changelog
<!-- Is this a notable change that affects users? If so, add a line to
`CHANGELOG.md` and prefix the line with one of the following:
    - [Added] for new features.
    - [Changed] for changes in existing functionality.
    - [Deprecated] for soon-to-be removed features.
    - [Removed] for now removed features.
    - [Fixed] for any bug fixes.
    - [Security] in case of vulnerabilities.
-->

[Changed] Link is now available as a payment method even if you collect
billing details via `billingDetailsCollectionConfiguration`.