Remove Futures / Promise based API client in favor of async client (#4757)
Context:
https://docs.google.com/document/d/1q8sSaEKv0AO_RXwpgbhtAkGL77NadeMf-Ab-UJuNa5U/edit?usp=sharing

## Summary

This removes the `FinancialConnectionsAPIClient` (Futures / Promise
based) in favor of the `FinancialConnectionsAsyncAPIClient`. Here's
where this leaves us:

- Our only API client is powered by Swift Concurrency.
- The API client conforms to the Futures / Promise based protocol.
- All call sites (view controllers / data sources) interface with a
Futures / Promise based API, as they already do today.

With that:
- We only maintain one API client going forward.
- We can still migrate all call sites to a Swift Concurrency based API
if we chose to.

## Motivation

Cleanup our in progress migration to Swift Concurrency.

## Testing

All of our E2E test have already been using the Async API client! I also
added more tests for the async poller in this PR.

## Changelog

N/a