Support CBC in native Link (#4693)
## Summary
<!-- Simple summary of what was changed. -->

This pull request expands support for card brand choice in the native
Link UI, allowing consumers to change the brand for existing cards.
Setting the desired brand for _new_ cards is already possible.

## Motivation
<!-- Why are you making this change? If it's for fixing a bug, if
possible, please include a code snippet or example project that
demonstrates the issue. -->

[LINK_MOBILE-81](https://jira.corp.stripe.com/browse/LINK_MOBILE-81)

## Testing
<!-- How was the code tested? Be as specific as possible. -->

Screenshot test.

## Changelog
<!-- Is this a notable change that affects users? If so, add a line to
`CHANGELOG.md` and prefix the line with one of the following:
    - [Added] for new features.
    - [Changed] for changes in existing functionality.
    - [Deprecated] for soon-to-be removed features.
    - [Removed] for now removed features.
    - [Fixed] for any bug fixes.
    - [Security] in case of vulnerabilities.
-->