Add exposure logging framework in MPE for Link Global Holdback (#4768)
## Summary

This implements a small exposure logging framework in PaymentSheet,
which logs the assignment of certain experiments such that their impact
can be observed. Importantly, these events are logged to `r.stripe.com`,
as opposed to other events in MPE which log to `q.stripe.com`. Because
of this, I've had to add an instance of `AnalyticsClientV2` in
`PaymentSheetAnalyticsHelper` that use the `client_id` and `origin`
values defined
[here](https://stripe.sourcegraphcloud.com/stripe-internal/zoolander/-/blob/src/resources/com/stripe/dscore/analyticseventlogger/server/rpcserver/client_config.yaml?L1667-1671)
in Zoolander.

This PR also makes use of that framework for the Link global holdback
experiment, which we log after the payment sheet session is fetched. In
order to do this, I've had to add a few fields to
STPElementsSession-related models, and have added Github comments with
their backend source.

More context in this doc:
https://docs.google.com/document/d/1hLJqa07kO6iv6jgVPJodZH240IK1e3iqW4rFa_7Wyc0/edit?usp=sharing

## Motivation

Enabling experiments in MPE!

## Testing

Unit tests added!

## Changelog

N/a