Fix another theming issue with Link verification fields (#4796)
## Summary
<!-- Simple summary of what was changed. -->

This pull request fixes a new theming issue in the Link verification
view, where the individual text fields would render in light mode even
if the device was in dark mode.

This was likely introduced in
https://github.com/stripe/stripe-ios/pull/4785, where I attempted to
correctly render the text fields if merchants set `alwaysLight` or
`alwaysDark` on MPE.

## Motivation
<!-- Why are you making this change? If it's for fixing a bug, if
possible, please include a code snippet or example project that
demonstrates the issue. -->

## Testing
<!-- How was the code tested? Be as specific as possible. -->

## Changelog
<!-- Is this a notable change that affects users? If so, add a line to
`CHANGELOG.md` and prefix the line with one of the following:
    - [Added] for new features.
    - [Changed] for changes in existing functionality.
    - [Deprecated] for soon-to-be removed features.
    - [Removed] for now removed features.
    - [Fixed] for any bug fixes.
    - [Security] in case of vulnerabilities.
-->