Fix Link selection issue in FlowController (#4793)
## Summary
<!-- Simple summary of what was changed. -->

This pull request fixes an issue in FlowController where selecting Link
after having previously selected an LPM wouldn't result in the payment
method being changed.

**Before**: Tapping `Pay with Link` after having filled out the card
form doesn't do anything. It also discards the selection if no new
selection is made.


https://github.com/user-attachments/assets/c5fff708-25e3-49f5-8ceb-9b3f383c39f5

**After**: Tapping `Pay with Link` after having filled out the card form
changes the payment method to `Link`. It also correctly preserves the
selection if no new selection is made.


https://github.com/user-attachments/assets/8e800e89-682d-433e-bbad-35ddcc18de0b

## Motivation
<!-- Why are you making this change? If it's for fixing a bug, if
possible, please include a code snippet or example project that
demonstrates the issue. -->

## Testing
<!-- How was the code tested? Be as specific as possible. -->

## Changelog
<!-- Is this a notable change that affects users? If so, add a line to
`CHANGELOG.md` and prefix the line with one of the following:
    - [Added] for new features.
    - [Changed] for changes in existing functionality.
    - [Deprecated] for soon-to-be removed features.
    - [Removed] for now removed features.
    - [Fixed] for any bug fixes.
    - [Security] in case of vulnerabilities.
-->