LPM rendering events (#4675)
## Summary
<!-- Simple summary of what was changed. -->
Added LPM render event to determine what the user sees without overflow.

For horizontal mode, the event has params "visible_lpms" and
"hidden_lpms" corresponding to the visible cells in the carousel and
wallet headers on render and the rest of the lpms that are not visible
in the carousel.
For vertical mode and embedded, the event has params "visible_lpms" and
"hidden_lpms", but there are no hidden lpms, as the user scrolls through
all of the lpms in order to pay.

Emitted on viewDidAppear for horizontal and vertical, emitted on layout
subviews for embedded.

LPM default selected (which LPM is default selected if any) can be found
on mc_form_shown - "selected_lpm"
LPM selected (which LPMs did buyer select if any) can be found on
mc_carousel_payment_method_tapped - "selected_lpm"
## Motivation
<!-- Why are you making this change? If it's for fixing a bug, if
possible, please include a code snippet or example project that
demonstrates the issue. -->
[MOBILESDK-3221](https://jira.corp.stripe.com/browse/MOBILESDK-3221)
## Testing
<!-- How was the code tested? Be as specific as possible. -->

- Unit test log
- UI tests that verify analytics

## Changelog
<!-- Is this a notable change that affects users? If so, add a line to
`CHANGELOG.md` and prefix the line with one of the following:
    - [Added] for new features.
    - [Changed] for changes in existing functionality.
    - [Deprecated] for soon-to-be removed features.
    - [Removed] for now removed features.
    - [Fixed] for any bug fixes.
    - [Security] in case of vulnerabilities.
-->
N/A