Complete cancelled FCLite sessions (#4813)
## Summary

When the FC Lite flow is dismissed by the user, we still attempt to get
a session object. In this scenario, we should call the `complete`
endpoint ourselves instead of the `session_receipt` endpoint. When the
flow is cancelled from within the webview (i.e. tapping the `X`), the
web flow calls the `complete` endpoint, so we just get the session
receipt.

## Motivation

Fixes bugs when swiping to dismiss.

## Testing

Before: 


https://github.com/user-attachments/assets/8e2500f6-2ebd-489b-9676-8f2d8cf3dc23

After:


https://github.com/user-attachments/assets/e1da957c-3a77-4be3-9e55-f23a6cb544da

Dismissing the flow during various points in the flow:


https://github.com/user-attachments/assets/42c4597a-f6e4-4ff5-8454-809079dd177e

## Changelog

N/a