Present from the main window VC if no VC is set in SwiftUI (#4827)
## Summary
In our Embedded + SwiftUI bindings, it's possible that the
`presentingViewController` has not yet been set when calling `confirm`.
For example, a user may have loaded embedded, displayed the
`currentPaymentOption`, but the user chose not to change it and
confirmed immediately.

In this case, we should set the `presentingViewController` on confirm,
as that's what we would do anyway once Embedded was shown. (We'll only
do this in SwiftUI for now, as that's the only situation where we manage
the `presentingViewController` for the user.

## Motivation
Fixing an assertion failure when paying without presenting the Embedded
VC

## Testing
Existing tests, tested in sample app

## Changelog
Minor fix, embedded not yet released