[LINK_MOBILE-122] Use nicknames for cards in Link (#4828)
Resolves
[LINK_MOBILE-122](https://jira.corp.stripe.com/browse/LINK_MOBILE-122)

## Summary

This moves the `••••\(last4)` label to the subtitle, and adds a nickname
to cards in the Link wallet. The nicknames either comes from the backend
(user manually set a nickname in app.link.com), or we construct a
reasonable default along the lines of `{brand} {funding}`.

This mimics the behavior on web. We'd love to have this entire logic
moved to the backend, and will work with that team to make those
changes.

## Motivation

Aligns with web

## Testing

| Web | Mobile |
|--------|--------|
| ![Screenshot 2025-04-22 at 10 16
16 AM](https://github.com/user-attachments/assets/c1f4d8fe-1aa5-4fba-9c1b-3810501cfb2f)
| ![Simulator Screenshot - iPhone 16 - 2025-04-22 at 11 17
51](https://github.com/user-attachments/assets/13cd9da9-ed35-4765-a471-ca51c1f61a16)
|

## Changelog

N/a