Add a flag and read from PMO SFU in PaymentSheetFormFactory (#4830)
## Summary
<!-- Simple summary of what was changed. -->
Renamed PaymentMethodOptionsSetupFutureUsageBeta to
PaymentMethodOptionsSetupFutureUsagePreview, added a
shouldReadPaymentMethodOptionsSetupFutureUsage to configurations and
checks PMO SFU in PaymentSheetFormFactory. This controls showing/hiding
the save checkbox and mandates.
## Motivation
<!-- Why are you making this change? If it's for fixing a bug, if
possible, please include a code snippet or example project that
demonstrates the issue. -->
PMO SFU project
[MOBILESDK-3406](https://jira.corp.stripe.com/browse/MOBILESDK-3406)
[MOBILESDK-3408](https://jira.corp.stripe.com/browse/MOBILESDK-3408)
## Testing
<!-- How was the code tested? Be as specific as possible. -->
Unit tests in PaymentSheetFormFactoryTests that check the visibility of
the save checkbox and mandates and the value of isSettingUp
## Changelog
<!-- Is this a notable change that affects users? If so, add a line to
`CHANGELOG.md` and prefix the line with one of the following:
    - [Added] for new features.
    - [Changed] for changes in existing functionality.
    - [Deprecated] for soon-to-be removed features.
    - [Removed] for now removed features.
    - [Fixed] for any bug fixes.
    - [Security] in case of vulnerabilities.
-->
N/A